+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 dd mm yyyy
 
+- Removed "invalid token" and "no token" lists from user admin page.
 - Add reason "moved-owner-changed" (used instead of "moved" if applicable) to "RemovedCharacter".
 - Small text improvements for the service admin page and better error handling if the service configuration is invalid.
 - Update dependencies.